cxr_v4/requirements.txt
python-multipart==0.0.5
opencv-python==4.5.5.64
onnxruntime==1.10.0
torch==1.9.0
torchvision==0.10.0
numpy==1.18.5
Pillow==8.1.2
requests==2.26.0
fastapi==0.63.0
uvicorn==0.13.4
requests-toolbelt==0.9.1